backend/requirements.txt
django>=5.2.1
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
pillow>=11.2.1
reportlab>=4.4.0
python-dotenv>=1.0.0
six>=1.16.0
django-rest-auth>=0.9.5
djangorestframework-simplejwt>=5.3.1
django-filter>=25.1
django-debug-toolbar
weasyprint
PyPDF2
django-chartjs
django-extensions
whitenoise
django-redis
openpyxl

# WebSocket support for real-time notifications
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0